+++ CodingInterview/CodingInterview/FindInPartiallySortedMatrix.cpp
@@ -1,12 +1,12 @@
 
-bool findInPartiallySortedMatrix(const int arr[], size_t rows, size_t columns, int number) {
+bool findInPartiallySortedMatrix(const int arr[], int rows, int columns, int number) {
 	if (arr == nullptr)
 		return false;
 
 	if (number < arr[0] || number > arr[rows * columns - 1])
 		return false;
 
-	size_t index = columns - 1;
+	int index = columns - 1;
 	while (index != (rows - 1) * columns) {
 		if (number < arr[index]) {
 			if (index % columns == 0)
